+++ app/views/items/index.html.erb
@@ -134,9 +134,11 @@
           <%= image_tag item.image, class: "item-img" %>
 
           <%# 商品が売れていればsold outを表示しましょう (こちらは購入機能に関する事なのでまだコメントアウトは残して起きたいです) %>
+          <% if item.purchase.present? %>
           <div class='sold-out'>
             <span>Sold Out!!</span>
           </div>
+          <% end %>
           <%# //商品が売れていればsold outを表示しましょう (こちらは購入機能に関する事なのでまだコメントアウトは残して起きたいです) %>
 
         </div>
